requirements.txt
langchain
langchain-community
langchain-ollama
langchain-postgres
psycopg[binary]
psycopg-pool
unstructured
markdown
python-dotenv
